Basic implementation of a buffer list
This commit adds a second bar under/above the status bar (according to the
`ncurses_status_on_top` UI option) which lists all the buffers opened. The
bar is rendered using the -now default- `BufList` face, and the display name
of the currently open buffer will be rendered using the -now default as well-
`BufListCurrent` face.

Menus take precedence over this buffer list, which means the bar will not be
displayed when an interactive selection is necessary (e.g. when completing
a command name and candidates show up), but other information boxes such
as the `autoinfo` will be rendered under the buffer bar.

This feature is disabled by default and has to be activated using the
`ncurses_buflist_show` UI option.